chatCNJ/chatCNJ/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.0
asyncpg==0.29.0
pgvector==0.3.2
openai==1.55.0
httpx==0.27.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
numpy==1.26.4
# The packages below should be pinned to specific versions for production.
# Pin them after verifying compatibility (examples shown but please adjust):
python-multipart==0.0.6
pypdf==3.13.0
python-docx==0.8.11
exa_py==1.0.0
